Fix: 댓글 평점 정수형 변경

+++ src/main/java/com/greedy/mokkoji/api/comment/dto/request/CommentCreateRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Min;
 public record CommentCreateRequest(
         @Min(0)
         @Max(5)
-        Double rate,
+        Integer rate,
         String content
 ) {
 }
